Fix glob pattern to match pdf file key names with a-z suffix
The existing glob for pdf files had a couple of errors.
The year prefix anchor had mixed case and the year only had three numbers.
The shell patterns for file name can't handle ? as optional
which was intended to be the alternate char for multiple papers
in the same year.
Fixed the glob and duplicated it to allow for the optional suffix.